+++ c/column_int.cc
@@ -23,8 +23,8 @@ SType IntColumn<T>::stype() const {
 }
 
 template <typename T>
-py::oobj IntColumn<T>::get_value_at_index(int64_t i) const {
-  int64_t j = (this->ri).nth(i);
+py::oobj IntColumn<T>::get_value_at_index(size_t i) const {
+  size_t j = (this->ri).nth(i);
   T x = this->elements_r()[j];
   return ISNA<T>(x)? py::None() : py::oint(x);
 }
